refactor CakeCard
