--- src/cakes/AddCake.js
@@ -1,13 +1,0 @@
-import React, { Component } from 'react'
-import PageWrapper from '../components/PageWrapper'
-
-
-export default class AddCake extends Component {
-    render() {
-        return (
-            <PageWrapper>
-                <h1>Dodaj ciasto do swojej oferty</h1>
-            </PageWrapper>
-        )
-    }
-}
