+++ src/cakes/CakeAddForm/RenderCakeAddform.js
@@ -5,7 +5,7 @@ import CakeAddInput from './component/CakeAddInput';
 import CakeAddTypesSelect from './component/CakeAddTypesSelect';
 import CakeAddSelect from './component/CakeAddSelect';
 import { YESNOSELECT } from '../../constans/selectConstans'
-import CookLabelFull from '../CookLabelFull';
+import CookLabelFull from '../cakeCard/CookLabelFull';
 
 function RenderCakeAddForm(props) {
 
@@ -137,7 +137,7 @@ function RenderCakeAddForm(props) {
                     powrót
                         </Button>
                 <Button onClick={props.addCakeFetch}
-                    variant="outlined"
+                    variant="outlined"cakeCard
                     color="primary"
                     style={{ margin: '20px' }}
                 >
